Fit the page content into a centered card.
Introduced a div container into index.html that frames all the page content.
Added a "recipe-card" class to which the div container belongs. Set style
for it in the styles.css file: white background, green border, centered in
the page (using "margin: 0 auto"), the frame has the width of the content
inside of it (using "width: fit-content").

To make the h1 work incorrectly, created a descendant combinator.

Also changed the font-weight and margin of the link selector "a" so the
list from index.page doesn't look like a continuous shape.